refactor: Car, CarService 클래스 분리 및 패키지 정리
