+++ src/main/java/racingcar/domain/Car.java
@@ -1,0 +1,11 @@
+package racingcar.domain;
+
+public class Car {
+    String carName;
+    int count;
+
+    public Car(String carName, int count) {
+        this.carName = carName;
+        this.count = count;
+    }
+}
